Create a graphic editor draft
Create a draft of a simple graphics editor of the paint type (needs to be cleaned and optimized).
Create stateless components for the graphical editor (toolbox, toolbox-settings, page-selector).
Create a sketch of the application logic based on redux.
Create functionality to allow the user to load images through the file dialog.

Change scss settings
Change eslint settings