Fix premature path closing.
Tool.ts
Swap the bodies of the finishDrawing and beginDrawing functions, because
finishDrawing uses ctx.beginPath() and beginDrawing uses ctx.closePath();

Tools.ts
Add a special tool type to mark the start of drawing in the history.

GraphicEditor.tsx
Enable canvas scaling with wheell move, as ctrl + wheell move is inconvenient for touchpad and graphics tablet users.

Add handleStartDrawing function to mark the start of drawing in history.
Add support for the start of drawing in the rerenderCanvas function.